+++ tests/test-dirs/locate/distinguish-files.t/bin/main.ml
@@ -1,0 +1,5 @@
+let _ = Lib_a.Same.foo
+let _ = Lib_b.Same.foo
+
+let _ = Lib_a.Different.foo
+let _ = Lib_b.Different.foo