+++ tests/test-dirs/locate/distinguish-files.t/lib_a/different.ml
@@ -1,0 +1,2 @@
+(* AAAAA *)
+let foo = 10