+++ tests/test-dirs/locate/distinguish-files.t/lib_b/different.ml
@@ -1,0 +1,2 @@
+(* BBBBB *)
+let foo = 10