+++ CHANGES.md
@@ -3,6 +3,8 @@ unreleased
 
   + merlin binary
     - Support for OCaml 5.3
+    - Use new 5.3 features to improve locate behavior in some cases. Merlin no
+      longer confuses uids from interfaces and implementations. (#1857)
   + vim plugin
     - Added support for search-by-type (#1846)
       This is exposed through the existing `:MerlinSearch` command, that