+++ CHANGES.md
@@ -6,6 +6,8 @@ unreleased
     - Use new 5.3 features to improve locate behavior in some cases. Merlin no
       longer confuses uids from interfaces and implementations. (#1857)
     - Perform less merges in the indexer (#1881)
+    - Add initial support for project-wide renaming: occurrences can now return
+      all usages of all related definitions. (#1877)
   + vim plugin
     - Added support for search-by-type (#1846)
       This is exposed through the existing `:MerlinSearch` command, that